+++ R/quickfit.R
@@ -11,9 +11,8 @@
 #' @param n.clust number of clusters (2 or 3)
 #' @param plotme whether or not to plot the result
 #' @return a named vector of initial estimates: 
-#' \itemize{
-#' \item{if \code{n.clust = 2}} {returns \code{t1, dt, x1, y1, x2, y2}} 
-#' \item{if \code{n.clust = 3}} {returns \code{t1, dt1, t2, dt2, x1, y1, x2, y2, x3, y3} }}
+#' - if `n.clust = 2`: returns `t1, dt, x1, y1, x2, y2`
+#' - if `n.clust = 3`: returns `t1, dt1, t2, dt2, x1, y1, x2, y2, x3, y3`
 #' @example ./demo/quickfit_example.r
 #' @export
 quickfit <-  function(T,X,Y, dt=1, n.clust = 2, plotme=TRUE){