+++ NEWS.md
@@ -1,3 +1,10 @@
-# marcher 0.0.2
+# marcher 0.0.3 (2026-09-14)
 
-* passed all of the R CMD check errors, warnings and notes
+- Minor fixes to documentation made to address CRAN check warnings and notes
+- 2020.04.27: Github version was updated to iron out some more issues, including an error in the range-shift testing function.
+- 2019.10.04: Github version was updated to iron out some issues (especially with plotting)
+- 2019.10.04: Github version was updated to iron out some issues (especially with plotting)
+
+# marcher 0.0.2 (2017-04-12)
+
+- passed all of the R CMD check errors, warnings and notes